Add tests that cover nested questions and models
