+++ e2e/test/scenarios/question/notebook-data-source.cy.spec.ts
@@ -1,6 +1,10 @@
 import { SAMPLE_DB_ID, WRITABLE_DB_ID } from "e2e/support/cypress_data";
 import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import { ORDERS_COUNT_QUESTION_ID } from "e2e/support/cypress_sample_instance_data";
+import {
+  SECOND_COLLECTION_ID,
+  ORDERS_COUNT_QUESTION_ID,
+  ORDERS_MODEL_ID,
+} from "e2e/support/cypress_sample_instance_data";
 import {
   popover,
   restore,
@@ -12,6 +16,8 @@ import {
   resyncDatabase,
   visualize,
   saveQuestion,
+  visitModel,
+  openQuestionActions,
 } from "e2e/support/helpers";
 const { REVIEWS_ID } = SAMPLE_DATABASE;
 
@@ -167,4 +173,141 @@ describe("scenarios > notebook > data source", () => {
       },
     );
   });
+
+  describe("saved entity as a source (aka the virtual table)", () => {
+    const modelDetails = {
+      name: "GUI Model",
+      query: { "source-table": REVIEWS_ID, limit: 1 },
+      display: "table",
+      type: "model",
+      collection_id: SECOND_COLLECTION_ID,
+    };
+
+    beforeEach(() => {
+      restore();
+      cy.signInAsAdmin();
+    });
+
+    it("data selector should properly show a model as the source (metabase#39699)", () => {
+      cy.createQuestion(
+        {
+          name: "GUI Model",
+          query: { "source-table": REVIEWS_ID, limit: 1 },
+          display: "table",
+          type: "model",
+          collection_id: SECOND_COLLECTION_ID,
+        },
+        { visitQuestion: true },
+      );
+      openNotebook();
+      cy.findByTestId("data-step-cell")
+        .should("have.text", modelDetails.name)
+        .click();
+
+      cy.findByTestId("saved-entity-back-navigation").should(
+        "have.text",
+        "Models",
+      );
+
+      cy.findByTestId("saved-entity-collection-tree").within(() => {
+        cy.findByLabelText("Our analytics")
+          .should("have.attr", "aria-expanded", "false")
+          .and("have.attr", "aria-selected", "false");
+        cy.findByLabelText("First collection")
+          .should("have.attr", "aria-expanded", "true")
+          .and("have.attr", "aria-selected", "false");
+        cy.findByLabelText("Second collection")
+          .should("have.attr", "aria-expanded", "false")
+          .and("have.attr", "aria-selected", "true");
+      });
+
+      cy.findByTestId("select-list")
+        .findByLabelText(modelDetails.name)
+        .should("have.attr", "aria-selected", "true");
+    });
+
+    it("moving the model to another collection should immediately be reflected in the data selector (metabase#39812-1)", () => {
+      visitModel(ORDERS_MODEL_ID);
+      openNotebook();
+
+      openDataSelector();
+      assertSourceCollection("Our analytics");
+      assertDataSource("Orders Model");
+
+      moveToCollection("First collection");
+
+      openDataSelector();
+      assertSourceCollection("First collection");
+      assertDataSource("Orders Model");
+    });
+
+    it("moving the source question should immediately reflect in the data selector for the nested question that depends on it (metabase#39812-2)", () => {
+      const SOURCE_QUESTION_ID = ORDERS_COUNT_QUESTION_ID;
+      // Rename the source question to make assertions more explicit
+      const sourceQuestionName = "Source Question";
+      cy.request("PUT", `/api/card/${ORDERS_COUNT_QUESTION_ID}`, {
+        name: sourceQuestionName,
+      });
+
+      const nestedQuestionDetails = {
+        name: "Nested Question",
+        query: { "source-table": `card__${SOURCE_QUESTION_ID}` },
+      };
+
+      cy.createQuestion(nestedQuestionDetails, {
+        wrapId: true,
+        idAlias: "nestedQuestionId",
+      });
+
+      visitQuestion("@nestedQuestionId");
+      openNotebook();
+
+      openDataSelector();
+      assertSourceCollection("Our analytics");
+      assertDataSource(sourceQuestionName);
+
+      cy.log("Move the source question to another collection");
+      visitQuestion(SOURCE_QUESTION_ID);
+      openNotebook();
+      moveToCollection("First collection");
+
+      cy.log("Make sure the source change is reflected in a nested question");
+      visitQuestion("@nestedQuestionId");
+      openNotebook();
+
+      openDataSelector();
+      assertSourceCollection("First collection");
+      assertDataSource(sourceQuestionName);
+    });
+  });
 });
+
+function moveToCollection(collection: string) {
+  openQuestionActions();
+  popover().findByTextEnsureVisible("Move").click();
+  cy.findByRole("dialog").within(() => {
+    cy.intercept("GET", "/api/collection/tree**").as("updateCollectionTree");
+    cy.findAllByTestId("item-picker-item")
+      .filter(`:contains(${collection})`)
+      .click();
+
+    cy.button("Move").click();
+    cy.wait("@updateCollectionTree");
+  });
+}
+
+function openDataSelector() {
+  cy.findByTestId("data-step-cell").click();
+}
+
+function assertItemSelected(item: string) {
+  cy.findByLabelText(item).should("have.attr", "aria-selected", "true");
+}
+
+function assertSourceCollection(collection: string) {
+  return assertItemSelected(collection);
+}
+
+function assertDataSource(questionOrModel: string) {
+  return assertItemSelected(questionOrModel);
+}
